# docs/swagger.yaml
basePath: /api/v1
definitions:
  controllers.Album:
    properties:
      artist:
        type: string
      price:
        type: number
      title:
        type: string
    type: object
  controllers.User:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: ingo.sauerzapf@gmail.com
    name: Ingo Sauerzapf
    url: https://linked.in/in/ingosauerzapf
  description: This is a sample server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Swagger Example API
  version: "1.0"
paths:
  /albums:
    post:
      consumes:
      - application/json
      description: add an album by posting json
      parameters:
      - description: Add model
        in: body
        name: album
        required: true
        schema:
          $ref: '#/definitions/controllers.Album'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.Album'
      summary: add an album
      tags:
      - albums
  /albums/{id}:
    get:
      consumes:
      - application/json
      description: Returns an album by it's id
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: Get an album by it's id
      tags:
      - albums
  /users:
    post:
      consumes:
      - application/json
      description: add a user with password to the system
      parameters:
      - description: Add model
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/controllers.User'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.User'
      summary: add a user
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      description: login user with email and password
      parameters:
      - description: Add model
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/controllers.User'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.User'
      summary: login user
      tags:
      - users
  /users/validate:
    get:
      consumes:
      - application/json
      description: validate user via his jwt token in cookie
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.User'
      summary: validate user
      tags:
      - users
swagger: "2.0"